Change DeprecationWarnings to logging.warning
DeprecationWarnings are hidden by default in python >= 2.7, so for them
to be usef we need to enable them. Doing so though enabled
DeprecationWarnings for every library and module you import, which
causes a lot of unrelated DeprecationWarnings to be printed. And you
can't do anything for most of them since they're from some other
library you're using.

Let's change these to logging.warning then and revert
DeprecationWarnings to their hidden default.