Rename SpanBuilder to Span and support 'ca' annotation. (#114)
* Rename SpanBuilder to Span and support 'ca' annotation.

I'm renaming and exposing py_zipkin.encoding.Span and
py_zipkin.encoding.Endpoint. The next step will be to have `with
zipkin_span` return the Span instance rather than the context manager
itself. That way it'll be possible to update the span fields directly
without need to have setters for all the properties.

```
with zipkin_span(...) as span:
    span.remote_endpoint = create_endpoint(service_name='mysql')
    span.tags.update({'key': 'value'})
```

As a side effect, with this change we also support the 'ca' binary
annotation.