Fix multithreading issues (#118)
* Fix multithreading issues

This is an alternative implementation for #111.

I didn't particularly like that everything was called *Storage which was
very confusing. Also having to subclass LocalStorage (or SpanStorage)
and override a property felt weird and error prone.
Also the suggested way of getting the current zipkin_attrs by calling
ThreadLocalStack().get() felt weird since why do I need to create a new
object every time?

This PR takes a different approach. Users will need to pass in a
`get_tracer` function, which will be invoked at runtime every time we
need to access the tracer. That makes it work in multithreaded
environments.
Also instead of calling it storage I called the new class Tracer. In the
future I'll probably attach also the transport handlers to it,
completely removing the need for the logging_context.

If users want to access the current context they can just call the
`get_zipkin_attrs()` tracer's function. To access the tracer it's enough
to use `get_default_tracer()` which will return the correct one.