Add protobuf encoding
This PR adds support for the protobuf v3 encoding.

I tested this by manually generating a few traces, writing them to a
file and sending them to zipkin with curl.

I'm checking in the zipkin.proto file. It's not really necessary but it
seemed to make sense to keep it next to the generated code.

I also changed `encoding.create_endpoint` to accept an `use_defaults`
param which disables guessing default values. That makes tests easier
since the guessed ips would change from host to host.