Add an extract_zipkin_attrs_from_headers() helper
It turns out that extracting B3 headers correctly is harder than it
sounds.  This helper function implements the standard as specified at:
https://github.com/openzipkin/b3-propagation

You can pass a dict or dict-like object to the new
py_zipkin.request_helpers.extract_zipkin_attrs_from_headers()
function, and get back a ZipkinAttrs instance you can trust.

In some cases, a sample_rate float and a use_128bit_trace_id flag are
required, so they appear in the function signature as well.

This should take care of single-header extraction, which is part of
Issue #98.

This may help with Issue #73 (missing or empty X-B3-Sampled).